Adjusts the data directory to include the storage version.

+++ JPPhotoManager/JPPhotoManager.Domain/AssetCatalog.cs
@@ -5,7 +5,7 @@ namespace JPPhotoManager.Domain
     public class AssetCatalog
     {
         public bool HasChanges { get; set; }
-        public double StorageVersion { get; set; }
+        public int StorageVersion { get; set; }
         public List<Folder> Folders { get; }
         public List<Asset> Assets { get; }
         public ImportNewAssetsConfiguration ImportNewAssetsConfiguration { get; set; }
